- for test event

+++ poller_kqueue.go
@@ -154,7 +154,7 @@ func (p *poller) readWrite(ev *syscall.Kevent_t) {
 	fd := int(ev.Ident)
 	c := p.getConn(fd)
 	if c != nil {
-		if ((ev.Filter & syscall.EV_ERROR) != 0) || ((ev.Filter & syscall.EV_EOF) != 0) {
+		if ((uint32(ev.Filter) & uint32(syscall.EV_ERROR)) != 0) || ((uint32(ev.Filter) & uint32(syscall.EV_EOF)) != 0) {
 			log.Printf("+++++ readWrite 111 xx: event error")
 			c.closeWithError(io.EOF)
 			return
